Added documentation for every class Possible changes:    1. pass true/false value for bias instead of pointer or keep both options    2. write constructors for all layers that accept as previousLayer both collection of layers and one layer (for simple networks)  TODO: Implement Base class for Neural Network with BFS algorithm implementation as it's feedForward and backpropagate methods

+++ tests/TestNeurons/NeuronTest.h
@@ -92,12 +92,12 @@ void testNeuronsXOR() {
                      << endl;
             }
 
-            for( int i = neurons.size() - 2; i >= 3; --i ) {
+            for( size_t i = neurons.size() - 2; i >= 3; --i ) {
                 neurons[i] -> calculateLoss();
             }
 
             /// backpropagate neurons
-            for( int i = neurons.size() - 1; i >= 3; --i ) {
+            for( size_t i = neurons.size() - 1; i >= 3; --i ) {
                 neurons[i] -> backpropagateNeuron();
             }
         }
@@ -106,7 +106,7 @@ void testNeuronsXOR() {
             cout << "Loss #" << iteration << ": " << loss / batchSize << endl;
 
         /// update weights
-        for (int i = neurons.size() - 1; i >= 3; --i) {
+        for (size_t i = neurons.size() - 1; i >= 3; --i) {
             neurons[i] -> updateWeights( learningRate, batchSize );
         }
     }
